pyproject.toml
[project]
name = "house-and-land-packager"
version = "0.1.0"
description = "House & Land Package Pricing, Clash Management, and Land Data Aggregation Platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "pydantic[email]>=2.0",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "bcrypt<4.1",  # passlib 1.7.4 is incompatible with bcrypt 4.1+
    "python-multipart>=0.0.9",
    "openpyxl>=3.1",
    "apscheduler>=3.10",
    "structlog>=24.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
agents = [
    "pdfplumber>=0.11",
    "anthropic>=0.40",
    "playwright>=1.45",
    "pandas>=2.2",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.6",
    "mypy>=1.11",
    "black>=24.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
